+++ package.json
@@ -43,6 +43,7 @@
     "@antfu/eslint-config": "^2.6.1",
     "@clack/prompts": "^0.7.0",
     "@egoist/tailwindcss-icons": "^1.7.1",
+    "@floating-ui/vue": "^1.0.3",
     "@iconify-json/ph": "^1.1.9",
     "@oku-ui/accordion": "workspace:^",
     "@oku-ui/alert-dialog": "workspace:^",
