+++ packages/vue/src/popper/PopperArrow.vue
@@ -1,0 +1,82 @@
+<script lang="ts">
+// import type { ScopePopper } from './Popper.vue'
+import type { ArrowProps } from '@oku-ui/arrow'
+import type { Side } from './utils'
+
+export interface PopperArrowProps extends ArrowProps {
+  scopeOkuPopper?: any
+}
+
+export const OPPOSITE_SIDE: Record<Side, Side> = {
+  top: 'bottom',
+  right: 'left',
+  bottom: 'top',
+  left: 'right',
+}
+</script>
+
+<script setup lang="ts">
+import { computed, defineOptions, defineProps } from 'vue'
+import { useForwardRef } from '@oku-ui/use-composable'
+import { OkuArrow } from '@oku-ui/arrow'
+
+import { usePopperContentInject } from './PopperContent.vue'
+
+defineOptions({
+  name: 'OkuPopperArrow',
+  inheritAttrs: false,
+})
+
+const props = defineProps<PopperArrowProps>()
+
+const forwardedRef = useForwardRef()
+
+const contentInject = usePopperContentInject('OkuPopperContent', props.scopeOkuPopper)
+const baseSide = computed(() => {
+  return contentInject?.placedSide.value ? OPPOSITE_SIDE[contentInject.placedSide.value] : ''
+})
+</script>
+
+<template>
+  <span
+    :ref="(el: any) => {
+      contentInject.onArrowChange(el)
+      return undefined
+    }"
+    :style="{
+      position: 'absolute',
+      left: contentInject.arrowX?.value ? `${contentInject.arrowX?.value}px` : undefined,
+      top: contentInject.arrowY?.value ? `${contentInject.arrowY?.value}px` : undefined,
+      [baseSide]: '0px',
+      transformOrigin: {
+        top: '',
+        right: '0 0',
+        bottom: 'center 0',
+        left: '100% 0',
+      }[contentInject.placedSide.value!],
+      transform: {
+        top: 'translateY(100%)',
+        right: 'translateY(50%) rotate(90deg) translateX(-50%)',
+        bottom: 'rotate(180deg)',
+        left: 'translateY(50%) rotate(-90deg) translateX(50%)',
+      }[contentInject.placedSide.value!],
+      visibility: contentInject.shouldHideArrow.value
+        ? 'hidden'
+        : undefined,
+    }"
+    v-bind="$attrs"
+  >
+    <OkuArrow
+      :is="is"
+      :ref="forwardedRef"
+      :as-child="asChild"
+      :style="{
+        ...$attrs.style as any,
+        // ensures the element can be measured correctly (mostly for if SVG)
+        display: 'block',
+      }"
+    >
+      <slot />
+    </OkuArrow>
+  </span>
+</template>
